Big Richard update coming through.
Hey hey hey, it's that scripting system you asked for! (Closes #46) The
code is two fuzzy equality operators short of a war crime, but it's here!
Most parts of it have tests, including every part of the standard library.
The standard library is in RichardFunctions.cs, but most of it is just
list functions and math functions. There are a lot of math functions.

But no Pow function. I think I missed something.