Add function for channel length
The [len] function allows the user to retrieve the length, in
characters, of any existing channel, printing it to the active
output(s). This can be useful for determining the length of specific
areas of output that vary in size, allowing the user to later make
modifications to specific sections of output in later postprocessing.